mongo分片集群脚本

+++ tools/soft/mongodb/create_simple_shard_cluster.sh
@@ -31,9 +31,15 @@ INSTALL_PATH=/root/mongo_install
 # 最终复制集工作目录
 REPLISET_INSTALL_PATH=/root/Mongo_${DB_NAME}_ReplisetWorkPlace/
 
-cd ${WORK_PATH} && rm -rf ${WORK_PATH}
-cd ${INSTALL_PATH} && rm -rf ${INSTALL_PATH}
-cd ${REPLISET_INSTALL_PATH} && rm -rf ${REPLISET_INSTALL_PATH}
+if [ -e "${WORK_PATH}" ]; then
+    cd ${WORK_PATH} && rm -rf ${WORK_PATH}
+fi
+if [ -e "${INSTALL_PATH}" ]; then
+    cd ${INSTALL_PATH} && rm -rf ${INSTALL_PATH}
+fi
+if [ -e "${REPLISET_INSTALL_PATH}" ]; then
+    cd ${REPLISET_INSTALL_PATH} && rm -rf ${REPLISET_INSTALL_PATH}
+fi
 
 echo "${IP_LIST_FILE}, ${TARGET_USER}, ${TARGET_PWD}, ${DB_NAME}, ${RS_NAME}, ${WORK_PATH}, ${INSTALL_PATH}, ${REPLISET_INSTALL_PATH}"
